+++ warehouse/upload.go
@@ -1000,8 +1000,10 @@ func (job *UploadJob) loadTable(tName string) (bool, error) {
 	alteredSchema, err := job.updateSchema(tName)
 	if err != nil {
 		status := model.TableUploadUpdatingSchemaFailed
+		errorsString := misc.QuoteLiteral(err.Error())
 		_ = job.tableUploadsRepo.Set(context.TODO(), job.upload.ID, tName, repo.TableUploadSetOptions{
 			Status: &status,
+			Error:  &errorsString,
 		})
 		return alteredSchema, fmt.Errorf("update schema: %w", err)
 	}
@@ -1050,8 +1052,10 @@ func (job *UploadJob) loadTable(tName string) (bool, error) {
 	err = job.whManager.LoadTable(tName)
 	if err != nil {
 		status := model.TableUploadExportingFailed
+		errorsString := misc.QuoteLiteral(err.Error())
 		_ = job.tableUploadsRepo.Set(context.TODO(), job.upload.ID, tName, repo.TableUploadSetOptions{
 			Status: &status,
+			Error:  &errorsString,
 		})
 		return alteredSchema, fmt.Errorf("load table: %w", err)
 	}
@@ -1817,12 +1821,20 @@ func (job *UploadJob) areIdentityTablesLoadFilesGenerated() (bool, error) {
 	var (
 		mergeRulesTable = warehouseutils.ToProviderCase(job.warehouse.Type, warehouseutils.IdentityMergeRulesTable)
 		mappingsTable   = warehouseutils.ToProviderCase(job.warehouse.Type, warehouseutils.IdentityMappingsTable)
+		tu              model.TableUpload
+		err             error
 	)
 
-	if tu, err := job.tableUploadsRepo.GetByUploadIDAndTableName(context.TODO(), job.upload.ID, mergeRulesTable); err != nil && tu.Location == "" {
+	if tu, err = job.tableUploadsRepo.GetByUploadIDAndTableName(context.TODO(), job.upload.ID, mergeRulesTable); err != nil {
+		return false, fmt.Errorf("table upload not found for merge rules table: %w", err)
+	}
+	if tu.Location == "" {
 		return false, fmt.Errorf("merge rules location not found: %w", err)
 	}
-	if tu, err := job.tableUploadsRepo.GetByUploadIDAndTableName(context.TODO(), job.upload.ID, mappingsTable); err != nil && tu.Location == "" {
+	if tu, err = job.tableUploadsRepo.GetByUploadIDAndTableName(context.TODO(), job.upload.ID, mappingsTable); err != nil {
+		return false, fmt.Errorf("table upload not found for mappings table: %w", err)
+	}
+	if tu.Location == "" {
 		return false, fmt.Errorf("mappings location not found: %w", err)
 	}
 	return true, nil